Create useUserProfile hook
